re #9 Fixed two cases in tests_svc where the unused attribute was still being passed a value.
re #34 updated name attributes to match between Session and SessionTemplate models. Also removed the automatic application of most template values to a new session. The system was never meant to work that way, but instead, the UI should present the template values as defaults and allow the user to change them. If the UI implementation sees it fit to skip this part, it is welcome to do so, but let's not assume it wants to.

removed a few stray product line attributes being passed in event creation, which somehow slipped back into the test suite recently. This requirement was removed in issue #6