getter/setters are now class attributes
A metaclass has been added to ObjectManager to simplify adding
getters and setters in subclasses. Each subclass of object manager will
define GETTERS and SETTERS with the attributes that are unique to
the subclass. When the class is created they will be merged with the
getters/setters defined on the base classes.

If a subclass maps a getter or setter attribute to None it will be
removed from the dictionary upon class creation. This is a simple way to
override known attributes used in base classes.